Rename "internal" modules, split adapted_dataset
