+++ mila_datamodules/vision/datasets/datasets_test.py
@@ -26,7 +26,7 @@ from mila_datamodules.registry import (
     locate_dataset_root_on_cluster,
 )
 from mila_datamodules.vision.coco_test import coco_required
-from mila_datamodules.vision.datasets import binary_mnist, mnist
+from mila_datamodules.vision.datasets import _binary_mnist, _mnist
 
 from .base_test import (
     DownloadableDatasetTests,
@@ -156,11 +156,11 @@ class TestCocoCaptions(VisionDatasetTests[tvd.CocoCaptions]):
 
 # NOTE: Here the 'original class' is already in `mila_datamodules.datasets.binary_mnist` because
 # we include fixes for bugs in the base class (nothing to do with the clusters though)
-class TestBinaryMNIST(ReadFromRoot[binary_mnist.BinaryMNIST]):
+class TestBinaryMNIST(ReadFromRoot[_binary_mnist.BinaryMNIST]):
     pass
 
 
-class TestBinaryEMNIST(ReadFromRoot[binary_mnist.BinaryEMNIST]):
+class TestBinaryEMNIST(ReadFromRoot[_binary_mnist.BinaryEMNIST]):
     @pytest.fixture(params=["byclass", "bymerge"])
     def split(self, request: _FixtureRequest[str]) -> str:
         return request.param
@@ -188,7 +188,7 @@ class TestCelebA(LoadsFromArchives[tvd.CelebA], DownloadableDatasetTests):
 
 
 # Same here for `MNIST`, we have a 'patch' that fixes an issue with dataset folder names on Beluga.
-class TestMNIST(ReadFromRoot[mnist.MNIST], DownloadableDatasetTests, Required):
+class TestMNIST(ReadFromRoot[_mnist.MNIST], DownloadableDatasetTests, Required):
     pass
